+++ lib/src/ast_visiting_suggestor.dart
@@ -2,9 +2,12 @@ import 'dart:async';
 
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:codemod/src/file_context.dart';
+import 'package:logging/logging.dart';
 
 import 'patch.dart';
 import 'suggestor.dart';
+
+final _log = Logger('AstVisitingSuggestor');
 
 /// Mixin that implements the [Suggestor] interface and makes it easier to write
 /// suggestors that operate as an [AstVisitor].
@@ -47,21 +50,38 @@ mixin AstVisitingSuggestor<R> on AstVisitor<R> {
   final _patches = <Patch>{};
 
   /// The context helper for the file currently being visited.
-  FileContext get context => _context;
-  FileContext _context;
+  FileContext get context {
+    if (_context != null) return _context!;
+    throw StateError('context accessed outside of a visiting context. '
+        'Ensure that your suggestor only accesses `this.context` inside an AST visitor method.');
+  }
+
+  FileContext? _context;
 
   Stream<Patch> call(FileContext context) async* {
     if (shouldSkip(context)) return;
 
-    final unit = shouldResolveAst(context)
-        ? (await context.getResolvedUnit()).unit
-        : context.getUnresolvedUnit();
+    CompilationUnit unit;
+    if (shouldResolveAst(context)) {
+      var result = await context.getResolvedUnit();
+      if (result == null || result.unit == null) {
+        _log.warning(
+            'Could not get resolved unit for "${context.relativePath}"');
+        return;
+      }
+      unit = result.unit!;
+    } else {
+      unit = context.getUnresolvedUnit();
+    }
+
     _patches.clear();
     _context = context;
     unit.accept(this);
     // Force the copying of this list, otherwise it would be a lazy iterable
     // mapped to the field on this class that will change on the next call.
     final patches = _patches.toList();
+    _context = null;
+
     yield* Stream.fromIterable(patches);
   }
 
@@ -76,7 +96,7 @@ mixin AstVisitingSuggestor<R> on AstVisitor<R> {
   /// contents if needed.
   bool shouldSkip(FileContext context) => false;
 
-  void yieldPatch(String updatedText, int startOffset, [int endOffset]) {
+  void yieldPatch(String updatedText, int startOffset, [int? endOffset]) {
     _patches.add(Patch(updatedText, startOffset, endOffset));
   }
 }